requirements.txt
transformers
datasets
tokenizers
scipy
torch
openai
backoff
pyarrow
python-dotenv